Unify VIES validation service
