+++ backend/shared/vies/src/ViesService.ts
@@ -1,14 +1,16 @@
 import type { AutoEncoderPatchType } from '@simonbackx/simple-encoding';
 import { isSimpleError, isSimpleErrors, SimpleError } from '@simonbackx/simple-errors';
+import { ViesCachedResult } from '@stamhoofd/models/models/ViesCachedResult.js';
 import type { Company } from '@stamhoofd/structures';
 import { PeppolScheme } from '@stamhoofd/structures';
 import { Country } from '@stamhoofd/types/Country';
 import jsvat from 'jsvat-next';
 import { PeppolDirectoryService } from './PeppolDirectoryService.js';
-import { ViesService } from './index.js';
+import { ViesApi } from './ViesApi.js';
 
-export class ViesHelperStatic {
-    testMode = false;
+const CacheMonths = 6;
+
+export class ViesServiceStatic {
 
     async checkCompany(company: Company, patch: AutoEncoderPatchType<Company> | Company) {
         // Validate the custom PEPPOL endpoint id, but only when it actually changed.
@@ -66,10 +68,11 @@ export class ViesHelperStatic {
 
         if (company.VATNumber !== null) {
             // Changed VAT number
-            patch.VATNumber = await ViesHelper.checkVATNumber(company.address.country, company.VATNumber);
+            const VATNumber = await this.checkVATNumber(company.address.country, company.VATNumber);
+            patch.VATNumber = VATNumber;
 
             if (company.address.country === Country.Belgium) {
-                patch.companyNumber = company.VATNumber.substring(2);
+                patch.companyNumber = VATNumber.substring(2);
             }
         }
 
@@ -78,7 +81,7 @@ export class ViesHelperStatic {
                 // Already validated
             } else {
                 // Need to validate
-                const result = await ViesHelper.checkCompanyNumber(company.address.country, company.companyNumber);
+                const result = await this.checkCompanyNumber(company.address.country, company.companyNumber);
                 patch.companyNumber = result.companyNumber;
                 if (result.VATNumber !== undefined) {
                     patch.VATNumber = result.VATNumber;
@@ -124,7 +127,7 @@ export class ViesHelperStatic {
                 VATNumber: corrected,
             };
         } catch (e) {
-            if (isSimpleError(e) || isSimpleErrors(e)) {
+            if ((isSimpleError(e) || isSimpleErrors(e)) && e.hasCode('invalid_field')) {
                 // Ignore: normal that it is not a valid VAT number
             } else {
                 // Other errors should be thrown
@@ -158,7 +161,7 @@ export class ViesHelperStatic {
         }
 
         try {
-            if (!await ViesService.checkVATNumber(country, formatted)) {
+            if (!await this.isVATNumberValid(country, formatted)) {
                 throw new SimpleError({
                     code: 'invalid_field',
                     message: $t('%1TG', { 'vat-number': formatted }),
@@ -179,6 +182,32 @@ export class ViesHelperStatic {
 
         return formatted;
     }
+
+    private async isVATNumberValid(country: Country, VATNumber: string): Promise<boolean> {
+        const cached = await ViesCachedResult.getByID(VATNumber);
+
+        const freshAfter = new Date();
+        freshAfter.setMonth(freshAfter.getMonth() - CacheMonths);
+
+        if (cached && cached.checkedAt >= freshAfter) {
+            return cached.result;
+        }
+
+        let result: boolean;
+        try {
+            result = await ViesApi.checkVATNumber(country, VATNumber);
+        }
+        catch (error) {
+            console.error('VIES error', error);
+            if (cached) {
+                return cached.result;
+            }
+            throw error;
+        }
+
+        await ViesCachedResult.saveResult(VATNumber, result, new Date());
+        return result;
+    }
 }
 
-export const ViesHelper = new ViesHelperStatic();
+export const ViesService = new ViesServiceStatic();
